+++ app/components/ahoy_captain/tables/headers/goals_header_component.rb
@@ -1,0 +1,9 @@
+module AhoyCaptain
+  module Tables
+    module Headers
+      class GoalsHeaderComponent < ViewComponent::Base
+
+      end
+    end
+  end
+end