+++ app/components/ahoy_captain/tables/rows/goals_row_component.rb
@@ -1,0 +1,12 @@
+module AhoyCaptain
+  module Tables
+    module Rows
+      class GoalsRowComponent < RowComponent
+
+        def total
+          @item.total_count
+        end
+      end
+    end
+  end
+end